download am and pm

+++ esmvaltool/cmorizers/data/downloaders/datasets/esacci_cloud.py
@@ -1,4 +1,5 @@
 """Script to download ESACCI-CLOUD."""
+import glob
 import logging
 import os
 
@@ -43,7 +44,12 @@ def download_dataset(config, dataset, dataset_info, start_date, end_date,
         overwrite=overwrite,
     )
 
-    base_path = 'https://public.satproj.klima.dwd.de/data/ESA_Cloud_CCI/CLD_PRODUCTS/v3.0/L3U/AVHRR-PM/'
+    base_path = 'https://public.satproj.klima.dwd.de/data/ESA_Cloud_CCI/CLD_PRODUCTS/v3.0/L3U/'
+
+    files = [
+        "*-ESACCI-L3U_CLOUD-CLD_MASKTYPE-AVHRR_*-fv3.0.nc",
+        "*-ESACCI-L3U_CLOUD-CLD_PRODUCTS-AVHRR_*-fv3.0.nc"
+    ]    
 
     wget_options = ['-r', '-nH', '-e', 'robots=off', '--cut-dirs=9', '--no-parent', '--reject="index.html*"']
 
@@ -52,15 +58,37 @@ def download_dataset(config, dataset, dataset_info, start_date, end_date,
         month = loop_date.month
         date = f'{year}{month:02}'
         if int(date) in range(200001, 200104):
-            sat = 'AVHRR_NOAA-14/'
-        elif int(date) in range(200104, 200509):
-            sat = 'AVHRR_NOAA-16/'
-        elif int(date) in range(200509, 200906):
-            sat = 'AVHRR_NOAA-18/'
+            sat_am = ''
+            sat_pm = 'AVHRR-PM/AVHRR_NOAA-14/'
+        elif int(date) in range(200104, 200211):
+            sat_am = ''
+            sat_pm = 'AVHRR-PM/AVHRR_NOAA-16/'
+        elif int(date) in range(200211, 200509):
+            sat_am = 'AVHRR-AM/AVHRR_NOAA-17/'
+            sat_pm = 'AVHRR-PM/AVHRR_NOAA-16/'
+        elif int(date) in range(200509, 200707):
+            sat_am = 'AVHRR-AM/AVHRR_NOAA-17/'
+            sat_pm = 'AVHRR-PM/AVHRR_NOAA-18/'
+        elif int(date) in range(200707, 200906):
+            sat_am = 'AVHRR-AM/AVHRR_METOPA/'
+            sat_pm = 'AVHRR-PM/AVHRR_NOAA-18/'
         else:
             logger.error("Number of instrument is not defined for date %s", date)
-        folder = base_path + sat + f'{year}/{month:02}'
-        logger.info("Download folder %s", folder) 
-        downloader.download_folder(folder, wget_options)
-        os.remove(os.path.join(downloader.local_folder, f'{month:02}'))
+        
+        for sat in (sat_am, sat_pm):
+            logger.info("sat = %s", sat)
+            if sat != '':
+                folder = base_path + sat + f'{year}/{month:02}'
+                logger.info("Download folder %s", folder)
+                #for file in files:
+                    #file_pattern = os.path.join(folder, file)
+                    #logger.info("filepattern =  %s", file_pattern)
+                    #dayfiles = sorted(glob.glob(file_pattern))
+                    #logger.info("Download files %s", dayfiles)
+                    #for filename in dayfiles:
+                    #logger.info("Download file %s", file_pattern) 
+                #downloader.download_file(folder, wget_options=['--accept="*-ESACCI-L3U_CLOUD-CLD_MASKTYPE-AVHRR_*-fv3.0.nc"'])
+                downloader.download_file(folder, wget_options)
+                #if os.path.exists(os.path.join(downloader.local_folder, f'{month:02}')):
+                os.remove(os.path.join(downloader.local_folder, f'{month:02}'))
         loop_date += relativedelta.relativedelta(months=1)
